Update to latest version and make less sweeping changes in server code
