+++ package.json
@@ -117,7 +117,7 @@
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
     "react-helmet": "^5.2.1",
-    "react-isomorphic-data": "^0.14.2",
+    "react-isomorphic-data": "^0.16.0",
     "react-lazyload": "^2.6.5",
     "react-router": "^5.1.2",
     "react-router-config": "^5.1.1",
