+++ src/server/index.jsx
@@ -1,14 +1,13 @@
-import React from 'react';
-import { renderToString, renderToStaticMarkup } from 'react-dom/server';
 import express from 'express';
 import compression from 'compression';
 import expressStaticGzip from 'express-static-gzip';
 import path from 'path';
 import helmet from 'helmet';
+
 // not part of react-helmet
-import { Helmet } from 'react-helmet';
 import gnuTP from 'gnu-terry-pratchett';
 import routes from '#app/routes';
+import renderDocument from './Document';
 import {
   articleDataPath,
   articleManifestPath,
@@ -27,22 +26,6 @@ import logResponseTime from './utilities/logResponseTime';
 import injectCspHeader, {
   localInjectHostCspHeader,
 } from './utilities/constructCspHeader';
-import Application from './Application';
-
-import {
-  DataProvider,
-  createDataClient,
-  getDataFromTree,
-} from 'react-isomorphic-data';
-import { renderToStringWithData } from 'react-isomorphic-data/ssr';
-
-import { StaticRouter } from 'react-router-dom';
-import SimorghApp from '../app/containers/App/App';
-import { ChunkExtractor } from '@loadable/server';
-import { ServerStyleSheet } from 'styled-components';
-import DocumentComponent from './Document/component';
-import getAssetOrigins from './utilities/getAssetOrigins';
-import { getStyleTag } from './styles';
 
 const fs = require('fs');
 
@@ -273,50 +256,6 @@ server
       });
     },
   )
-  .get('/test', async (req, res) => {
-    const dataClient = createDataClient({
-      initialCache: {},
-      ssr: true,
-    });
-
-    const reactApp = (
-      <DataProvider client={dataClient}>
-        <Application />
-      </DataProvider>
-    );
-
-    try {
-      await getDataFromTree(reactApp, dataClient);
-    } catch (err) {
-      console.log(`ohhhh ohh: ${err}`);
-    }
-
-    let markup;
-    // pass the same dataClient instance you are passing to your provider here
-    try {
-      markup = await renderToStringWithData(reactApp, dataClient);
-    } catch (err) {
-      console.error('Error while trying to getDataFromTree', err);
-    }
-
-    res.send(
-      `<!doctype html>
-          <html lang="">
-          <head>
-            <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-            <meta charSet='utf-8' />
-            <title>Razzle TypeScript</title>
-            <meta name="viewport" content="width=device-width, initial-scale=1">
-            <script>
-              window.__cache=${JSON.stringify(dataClient.cache)}
-            </script>
-          </head>
-          <body>
-            <div id="root">${markup}</div>
-          </body>
-        </html>`,
-    );
-  })
   .get('/*', cspInjectFun, async ({ url, headers, path: urlPath }, res) => {
     logger.info(
       JSON.stringify(
@@ -340,13 +279,7 @@ server
       data.path = urlPath;
       data.timeOnServer = Date.now();
 
-      // Create a store for all component data fetches
-      const dataClient = createDataClient({
-        initialCache: {},
-        ssr: true,
-      });
-
-      const ServerAppProps = {
+      const result = await renderDocument({
         bbcOrigin,
         data,
         isAmp,
@@ -354,102 +287,21 @@ server
         service,
         url,
         variant,
-      };
-
-      const ServerApp = ({ ServerAppProps }) => (
-        <DataProvider client={dataClient}>
-          <StaticRouter location={url} {...ServerAppProps}>
-            <SimorghApp
-              initialData={ServerAppProps.data}
-              routes={ServerAppProps.routes}
-              bbcOrigin={ServerAppProps.bbcOrigin}
-            />
-          </StaticRouter>
-        </DataProvider>
-      );
-
-      const sheet = new ServerStyleSheet();
-
-      const statsFile = path.resolve(
-        `${__dirname}/public/loadable-stats-${process.env.SIMORGH_APP_ENV}.json`,
-      );
-
-      const extractor = new ChunkExtractor({ statsFile });
-
-      const context = {};
-      let app;
-      try {
-        app = await renderToStringWithData(
-          extractor.collectChunks(
-            sheet.collectStyles(<ServerApp ServerAppProps={ServerAppProps} />),
-          ),
-          dataClient,
-        );
-      } catch (err) {
-        console.error(err);
-      }
-
-      try {
-        await getDataFromTree(app, dataClient);
-      } catch (err) {
-        console.log(`ohhhh ohh: ${err}`);
-      }
-
-      let markup;
-
-      const scripts = extractor.getScriptElements({
-        crossOrigin: 'anonymous',
-        type: 'text/javascript',
-        defer: true,
       });
-      const headHelmet = Helmet.renderStatic();
-      const assetOrigins = getAssetOrigins(service);
 
-      // pass the same dataClient instance you are passing to your provider here
-      try {
-        // markup = renderToString(app);
-        // markup = await renderToStringWithData(app, dataClient);
-        const doc = renderToStaticMarkup(
-          <DocumentComponent
-            assetOrigins={assetOrigins}
-            scripts={scripts}
-            app={app}
-            data={data}
-            componentData={dataClient.cache}
-            styleTags={getStyleTag(sheet, isAmp)}
-            helmet={headHelmet}
-            service={service}
-            isAmp={isAmp}
-          />,
-        );
-
-        res.status(status).send(`<!doctype html>${doc}`);
-      } catch (err) {
-        console.error('Error while trying to getDataFromTree', err);
+      if (result.redirectUrl) {
+        res.redirect(301, result.redirectUrl);
+      } else if (result.html) {
+        res.status(status).send(result.html);
+      } else {
+        throw new Error('unknown result');
       }
-
-      //   const result = await renderDocument({
-      //     bbcOrigin,
-      //     data,
-      //     isAmp,
-      //     routes,
-      //     service,
-      //     url,
-      //     variant,
-      //   });
-
-      //   if (result.redirectUrl) {
-      //     res.redirect(301, result.redirectUrl);
-      //   } else if (result.html) {
-      //     res.status(status).send(result.html);
-      //   } else {
-      //     throw new Error('unknown result');
-      //   }
     } catch ({ message, status }) {
       logger.error(
         JSON.stringify(
           {
             event: 'ssr_request_failed',
+            // ok. so status comes from getInitialData.
             status: status || 500,
             message,
             url,
